ml/feedback_generator/requirements.txt
# LLM Feedback Generator — Python Dependencies
# ------------------------------------------------

# API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# Database
supabase>=2.3.0
postgrest>=0.16.0

# HTTP client (for LLM APIs + supabase SDK)
httpx>=0.27.0

# Logging
loguru>=0.7.0

# Config
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Shared ML infrastructure (cache + prediction logging + Celery)
redis>=5.0.0
celery[redis]>=5.4.0
